Automatic reading of splices (no config section)
We used to specify the splice point and the name of the segment and
everything else about splices in the config file. Converting to a
different config system necessitated getting rid of that config section
dictionary. It was also a good time to generate the info dynamically
from files on disk and use alphabetical sorting to figure out the order
to apply the segments.